requirements.txt
gradio
PyPDF2
google-generativeai